Update repo_load_controller.py to use orm syntax.
