+++ augur/api/view/routes.py
@@ -74,17 +74,15 @@ def repo_table_view():
     pagination_offset = config.get_value("frontend", "pagination_offset")
     
     if current_user.is_authenticated:
-        data = load_repos_test(user = current_user, search = query, page = page, sort = sorting, direction = direction, source = "user")
-        page_count = load_repos_test(user = current_user, search = query, count = True, source = "user")
-        # data = current_user.get_repos(page = page, sort = sorting, direction = direction, search=query)[0]
-        # page_count = (current_user.get_repo_count(search = query)[0] or 0) // pagination_offset
+        data = current_user.get_repos(page = page, sort = sorting, direction = direction, search=query)[0]
+        repos_count = (current_user.get_repo_count(search = query)[0] or 0)
     else:
-        data = load_repos_test(search = query, page = page, sort = sorting, direction = direction)
-        page_count = load_repos_test(search = query, count = True)
-        # data = get_all_repos(page = page, sort = sorting, direction = direction, search=query)[0]
-        # page_count = (get_all_repos_count(search = query)[0] or 0) // pagination_offset
+        data = get_all_repos(page = page, sort = sorting, direction = direction, search=query)[0]
+        repos_count = (get_all_repos_count(search = query)[0] or 0)
+
+    page_count = math.ceil(repos_count / pagination_offset) - 1
     
-    if not data.count():
+    if not data:
         data = None
