+++ augur/api/view/utils.py
@@ -312,101 +312,4 @@ def render_module(module, **args):
     return render_template('index.j2', **args)
 
 """ ----------------------------------------------------------------
-    No longer used
 """
-# My attempt at a loading page
-def renderLoading(dest, query, request):
-    cache_files_requested.append(request)
-    return render_template('index.j2', body="loading", title="Loading", d=dest, query_key=query, api_url=getSetting('serving'))
-
-with DatabaseEngine() as engine:
-    augur_data_schema = MetaData(schema = "augur_data")
-    augur_data_schema.reflect(bind = engine, views = True)
-    
-    commits_materialized_view: Table = augur_data_schema.tables["augur_data.api_get_all_repos_commits"]
-    issues_materialized_view: Table = augur_data_schema.tables["augur_data.api_get_all_repos_issues"]
-
-""" ----------------------------------------------------------------
-"""
-def load_repos_test(count = False, source = None, **kwargs):
-    columns: list[Column] = [
-        Repo.repo_id.distinct().label("repo_id"),
-        Repo.description.label("description"),
-        Repo.repo_git.label("url"),
-        coalesce(commits_materialized_view.columns.commits_all_time, 0).label("commits_all_time"),
-        coalesce(issues_materialized_view.columns.issues_all_time, 0).label("issues_all_time"),
-        RepoGroup.rg_name.label("rg_name"),
-        Repo.repo_git.regexp_replace('.*github\.com\/[A-Za-z0-9 \- _]+\/([A-Za-z0-9 \- _ .]+)$', "\\1").label("repo_name"),
-        Repo.repo_git.regexp_replace('.*github\.com\/([A-Za-z0-9 \- _]+)\/[A-Za-z0-9 \- _ .]+$', "\\1").label("repo_owner"),
-        RepoGroup.repo_group_id.label("repo_group_id")
-    ]
-    
-    def get_colum_by_label(label: str)-> Column:
-        for column in columns:
-            if column.name == label:
-                return column
-    
-    repos: Query = db_session.query(*columns)\
-        .outerjoin(commits_materialized_view, Repo.repo_id == commits_materialized_view.columns.repo_id)\
-        .outerjoin(issues_materialized_view, Repo.repo_id == issues_materialized_view.columns.repo_id)\
-        .join(RepoGroup, Repo.repo_group_id == RepoGroup.repo_group_id)
-    
-    if source == "user":
-        user: User = kwargs.get("user")
-        
-        if not user:
-            return None, {"status": "User not passed when trying to get user repos"}
-        if not user.groups:
-            return None, {"status": "No data"}
-        
-        repos = repos.join(UserRepo, Repo.repo_id == UserRepo.repo_id)\
-            .join(UserGroup, UserGroup.group_id == UserRepo.group_id)\
-            .filter(UserGroup.user_id == user.user_id)
-    
-    elif source == "group":
-        user: User = kwargs.get("user")
-        
-        if not user:
-            return None, {"status": "User not specified"}
-        group_name = kwargs.get("group_name")
-        if not group_name:
-            return None, {"status": "Group name not specified"}
-        
-        group_id = ... # UserGroup.convert_group_name_to_id(self.session, user.user_id, group_name)
-        if group_id is None:
-            return None, {"status": "Group does not exists"}
-        
-        repos = repos.join(UserRepo, Repo.repo_id == UserRepo.repo_id)\
-            .filter(UserRepo.group_id == group_id)
-    
-    search = kwargs.get("search")
-    qkey = kwargs.get("query_key") or ["repo_name", "repo_owner"]
-    if search:
-        if isinstance(qkey, list) and len(qkey) > 0:
-            repos = repos.filter(or_(ilike_op(get_colum_by_label(filter_column), f"%{search}%") for filter_column in qkey))
-        else:
-            repos = repos.filter(ilike_op(get_colum_by_label(qkey), f"%{search}%"))
-    
-    page_size: int = kwargs.get("page_size") or 25
-    if count:
-        c = repos.count()
-        return math.ceil(c / page_size) - 1
-    else: 
-        page: int = kwargs.get("page") or 0
-        offset = page * page_size
-        direction = kwargs.get("direction") or "ASC"
-        order_by = kwargs.get("order_by") or "repo_id"
-        
-        if direction not in ["ASC", "DESC"]:
-            return None, None, {"status": "Invalid direction"}
-        
-        if order_by not in ["repo_id", "repo_name", "repo_owner", "commits_all_time", "issues_all_time"]:
-            return None, None, {"status": "Invalid order by"}
-        
-        # Find the column named in the 'order_by', and get its asc() or desc() method 
-        directive: function = getattr(get_colum_by_label(order_by), direction.lower())
-        
-        repos = repos.order_by(directive())
-    
-    return repos.slice(offset, offset + page_size)
-
